refactor: Split out `Immutable`
